requirements.txt
albumentations==0.5.2
braceexpand==0.1.7
easydict==1.9
imageio==2.5.0
imgaug==0.4.0
joblib==1.1.0
keras==2.7.0
kornia==0.5.0
matplotlib==3.4.3
memory_profiler==0.60.0
numpy==1.21.4
omegaconf==2.1.1
opencv_python==4.5.5.62
pandas==1.3.5
Pillow==9.1.0
pytorch_lightning==1.2.9
PyYAML==6.0
scikit_image==0.18.3
scikit_learn==1.0.2
scipy==1.7.2
six==1.16.0
tensorflow==2.7.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
webdataset==0.1.103
